feat: move cryptokey logic to a hook and remove socket connection from being in context
