+++ client/src/lib/decryptMessage.js
@@ -1,4 +1,5 @@
 export default async (message, cryptoKey) => {
+	console.log(cryptoKey)
 	const encryptedMessageUint8Array = new Uint8Array(
 		atob(message)
 			.split('')
